# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LiventCord",
    "version": "1.0"
  },
  "paths": {
...
    "/api/guilds": {
      "get": {
        "tags": [
          "Guild"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Guild"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "GuildName": {
                    "type": "string"
                  },
                  "region": {
                    "type": "string"
                  },
                  "Photo": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "GuildName": {
                  "style": "form"
                },
                "region": {
                  "style": "form"
                },
                "Photo": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/guilds/{guildId}/channels/{channelId}/writing": {
      "post": {
        "tags": [
          "Typing"
        ],
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}